+++ Sources/PayPalWebPayments/Environment+PayPalWebCheckout.swift
@@ -12,6 +12,13 @@ extension Environment {
             return URL(string: "https://www.sandbox.paypal.com")!
         case .live:
             return URL(string: "https://www.paypal.com")!
+        #if DEBUG
+        case .custom:
+            if let host = graphQLURL.host, let url = URL(string: "https://\(host)") {
+                return url
+            }
+            return URL(string: "https://www.paypal.com")!
+        #endif
         }
     }
 }